feat(messages): Add status attribute
This commit adds a `status` attribute to the `Messages` model. The `status` attribute can be one of three values: `Sending`, `Sent`, or `Failed`.

The `status` attribute is calculated based on the values of the `sent_at` and `failed_at` attributes. If neither `sent_at` nor `failed_at` is set, the `status` is `Sending`. If `sent_at` is set and `failed_at` is not set, the `status` is `Sent`. If `failed_at` is set and `sent_at` is not set, the `status` is `Failed`.

Signed-off-by: MarJose123 <18107626+MarJose123@users.noreply.github.com>