feat: remove hasConfigFile in ServiceProvider
This commit removes the hasConfigFile() method from the ServiceProvider class.

The hasConfigFile() method was used to determine if the service provider had a configuration file. However, this method is no longer necessary because the package discovery process now automatically detects the presence of a configuration file.

Removing the hasConfigFile() method simplifies the ServiceProvider class and makes it more consistent with other service providers in Laravel.

Signed-off-by: MarJose123 <18107626+MarJose123@users.noreply.github.com>